Move into ZarrCore.jl
- Move src/ to lib/ZarrCore/src/
- Move deps, establish workspace, etc.
- Change some names around to make Zarr.jl references into ZarrCore.jl references
