+++ lib/ZarrCore/Project.toml
@@ -1,0 +1,43 @@
+name = "Zarr"
+uuid = "77f5b75c-4c08-499f-ba13-550b0a0af171"
+version = "0.10.1"
+authors = ["Fabian Gans <fgans@bgc-jena.mpg.de>"]
+
+[deps]
+Blosc = "a74b3585-a348-5f62-a45c-50e91977d574"
+CRC32c = "8bf52ea8-c179-5cab-976a-9e18b702a9bc"
+ChunkCodecCore = "0b6fb165-00bc-4d37-ab8b-79f91016dbe1"
+ChunkCodecLibZlib = "4c0bbee4-addc-4d73-81a0-b6caacae83c8"
+ChunkCodecLibZstd = "55437552-ac27-4d47-9aa3-63184e8fd398"
+DataStructures = "864edb3b-99cc-5e75-8d2d-829cb0a9cfe8"
+DateTimes64 = "b342263e-b350-472a-b1a9-8dfd21b51589"
+Dates = "ade2ca70-3891-5945-98fb-dc099432e06a"
+DiskArrays = "3c3547ce-8d99-4f5e-a174-61eb10b00ae3"
+HTTP = "cd3eb016-35fb-5094-929b-558a96fad6f3"
+JSON = "682c06a0-de6a-54ab-a142-c8b1cf79cde6"
+OffsetArrays = "6fe1bfb0-de20-5000-8ca7-80f57d26f881"
+OpenSSL = "4d8831e6-92b7-49fb-bdf8-b643e874388c"
+OrderedCollections = "bac558e1-5e72-5ebc-8fee-abe8a469f55d"
+Pkg = "44cfe95a-1eb2-52ea-b672-e2afdf69b78f"
+URIs = "5c2747f8-b7ea-4ff2-ba2e-563bfd36b1d4"
+Unicode = "4ec0a83e-493e-50e2-b9ac-8f72acf5a8f5"
+ZipArchives = "49080126-0e18-4c2a-b176-c102e4b3760c"
+
+[compat]
+Blosc = "0.5, 0.6, 0.7"
+CRC32c = "1.10, 1.11"
+ChunkCodecCore = "1"
+ChunkCodecLibZlib = "1"
+ChunkCodecLibZstd = "1"
+DataStructures = "0.17, 0.18, 0.19"
+DateTimes64 = "1"
+DiskArrays = "0.4.21"
+HTTP = "^1.3, 2"
+JSON = "0.21, 1"
+OffsetArrays = "0.11, 1.0"
+OpenSSL = "1"
+OrderedCollections = "1.8.2, 2"
+URIs = "1"
+Unicode = "1.10, 1.11.0"
+ZipArchives = "2"
+julia = "1.10"
