+++ lib/ZarrCore/src/Filters/delta.jl
