+++ lib/ZarrCore/src/Filters/fixedscaleoffset.jl
