+++ lib/ZarrCore/src/Filters/quantize.jl
