+++ lib/ZarrCore/src/MaxLengthStrings.jl
