+++ lib/ZarrCore/src/Storage/consolidated.jl
