+++ lib/ZarrCore/src/Storage/directorystore.jl
