+++ lib/ZarrCore/src/ZarrCore.jl
@@ -1,0 +1,33 @@
+module ZarrCore
+
+import JSON
+import Blosc
+import Unicode
+using OrderedCollections: OrderedDict
+
+struct ZarrFormat{V}
+  version::Val{V}
+end
+Base.Int(v::ZarrFormat{V}) where V = V
+@inline ZarrFormat(v::Int) = ZarrFormat(Val(v))
+ZarrFormat(v::ZarrFormat) = v
+#Default Zarr Version
+const DV = ZarrFormat(Val(2))
+
+include("types.jl")
+include("chunkkeyencoding.jl")
+include("metadata.jl")
+include("metadata3.jl")
+include("Compressors/Compressors.jl")
+include("Codecs/Codecs.jl")
+include("Storage/Storage.jl")
+include("Filters/Filters.jl")
+include("ZArray.jl")
+include("pipeline.jl")
+include("ZGroup.jl")
+include("caching.jl")
+
+export ZArray, ZGroup, zopen, zzeros, zcreate, storagesize, storageratio,
+  zinfo, DirectoryStore, S3Store, GCStore, zgroup
+
+end # module
