+++ lib/ZarrCore/src/metadata3.jl
