+++ lib/ZarrCore/src/pipeline.jl
